+++ nautilus_trader/execution/engine.pyx
@@ -703,7 +703,7 @@ cdef class ExecutionEngine(Component):
 
     cdef void _update_position(self, Position position, OrderFilled fill) except *:
         # Check for flip
-        if position.is_opposite_side(fill.order_side) and fill.last_qty > position.quantity:
+        if position.is_opposite_side(fill.side) and fill.last_qty > position.quantity:
             self._flip_position(position, fill)
             return  # Handled in flip
 
@@ -759,7 +759,8 @@ cdef class ExecutionEngine(Component):
             venue_order_id=fill.venue_order_id,
             execution_id=fill.execution_id,
             position_id=fill.position_id,
-            order_side=fill.order_side,
+            order_side=fill.side,
+            order_type=fill.type,
             last_qty=position.quantity,  # Fill original position quantity remaining
             last_px=fill.last_px,
             currency=fill.currency,
@@ -789,7 +790,8 @@ cdef class ExecutionEngine(Component):
             venue_order_id=fill.venue_order_id,
             execution_id=fill.execution_id,
             position_id=position_id_flip,
-            order_side=fill.order_side,
+            order_side=fill.side,
+            order_type=fill.type,
             last_qty=difference,  # Fill difference from original as above
             last_px=fill.last_px,
             currency=fill.currency,