+++ nautilus_trader/model/orders/stop_limit.pyx
@@ -211,18 +211,18 @@ cdef class StopLimitOrder(PassiveOrder):
         Raises
         ------
         ValueError
-            If init.order_type is not equal to STOP_LIMIT.
+            If init.type is not equal to STOP_LIMIT.
 
         """
         Condition.not_none(init, "init")
-        Condition.equal(init.order_type, OrderType.STOP_LIMIT, "init.order_type", "OrderType")
+        Condition.equal(init.type, OrderType.STOP_LIMIT, "init.type", "OrderType")
 
         return StopLimitOrder(
             trader_id=init.trader_id,
             strategy_id=init.strategy_id,
             instrument_id=init.instrument_id,
             client_order_id=init.client_order_id,
-            order_side=init.order_side,
+            order_side=init.side,
             quantity=init.quantity,
             price=Price.from_str_c(init.options["price"]),
             trigger=Price.from_str_c(init.options["trigger"]),