+++ nautilus_trader/model/orders/stop_market.pyx
@@ -171,18 +171,18 @@ cdef class StopMarketOrder(PassiveOrder):
         Raises
         ------
         ValueError
-            If init.order_type is not equal to STOP_MARKET.
+            If init.type is not equal to STOP_MARKET.
 
         """
         Condition.not_none(init, "init")
-        Condition.equal(init.order_type, OrderType.STOP_MARKET, "init.order_type", "OrderType")
+        Condition.equal(init.type, OrderType.STOP_MARKET, "init.type", "OrderType")
 
         return StopMarketOrder(
             trader_id=init.trader_id,
             strategy_id=init.strategy_id,
             instrument_id=init.instrument_id,
             client_order_id=init.client_order_id,
-            order_side=init.order_side,
+            order_side=init.side,
             quantity=init.quantity,
             price=Price.from_str_c(init.options["price"]),
             time_in_force=init.time_in_force,